example/example_schema_plugin/pyproject.toml
[build-system]
requires = [ "setuptools>=61.0.0",]
build-backend = "setuptools.build_meta"

[project]
name = "example_schema"
version = "0.0.1"
description = "A plugin for NOMAD"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [ "Programming Language :: Python :: 3", "Programming Language :: Python :: 3 :: Only", "Programming Language :: Python :: 3.9",]
dependencies = [ "nomad-lab>=1.2.0-pre", "pytest", "typing-extensions==4.4.0",]

[project.license]
file = "LICENSE"

[tool.setuptools.packages.find]
where = [ "src",]
